refactor jobShow to its own file
move jobShow out of index.js to components/jobShow.js
