+++ lib/components/job.js
@@ -1,19 +1,15 @@
 import React from 'react';
 import Technology from './technology';
+import helpers from '../helpers';
 
 class Job extends React.Component {
   constructor() {
     super();
   }
 
-  //var createmarkup = () => {
-    //return {__html: fulllisting ? job.description : trimbody(job.description, 100)};
-                //<div className="dangerous" dangerouslySetInnerHTML={createMarkup()}></div>
-  //};
-                //{fullListing ? "" : <a className="btn btn-default" href="/#/job/1">View Details</a>}
-                //{fullListing ? <a href="/">Back to all jobs</a> : ""}
-                  //{this.props.job.raw_technologies.map((technology) => {
-                    //return (<li><Technology key={technology.id} technology={technology}/></li>)
+  createMarkup() {
+    return {__html: this.props.fullListing ? this.props.job.description : helpers.trimBody(this.props.job.description, 100)}
+    };
 
   render() {
     if (typeof this.props.job.title !== "undefined"){
@@ -28,9 +24,12 @@ class Job extends React.Component {
           <div className="row">
             <div className="col-md-6">
               <div className="content">
+                <div className="dangerous" dangerouslySetInnerHTML={this.createMarkup()}></div>
                 <a className="btn btn-default" target="_blank" href={this.props.job.url}>View job posting</a>
               </div> {/* <!-- content --> */}
-              <Technology technologies={this.props.job.raw_technologies} />
+              {/*<Technology technologies={this.props.job.raw_technologies} />*/}
+              {this.props.fullListing ? "" : <a className="btn btn-default" href="/#/job/1">View Details</a>}
+              {this.props.fullListing ? <a href="/">Back to all jobs</a> : ""}
             </div>
           </div>{/* <!-- row --> */}
         </div>
@@ -39,6 +38,6 @@ class Job extends React.Component {
       return <p>waiting</p>
     }
   }
-}
+};
 
 export default Job;
