+++ lib/components/jobListings.js
@@ -11,7 +11,7 @@ class JobListings extends React.Component {
       return(
         <div className="container">
         { this.props.jobs.map((job) => {
-          return <Job key={job.id} job={job} />;
+          return <Job key={job.id} job={job} fullListing={false} />;
         })}
         </div>
       );
